Refactor search -> search/core
