+++ sickbeard/properFinder.py
@@ -37,7 +37,7 @@ from sickrage.show.History import History
 from . import db, helpers, logger
 from .common import Quality, cpu_presets
 from .name_parser.parser import InvalidNameException, InvalidShowException, NameParser
-from .search import pickBestResult, snatchEpisode
+from .search.core import pickBestResult, snatchEpisode
 
 
 class ProperFinder(object):  # pylint: disable=too-few-public-methods
